OBPIH-3642 Improve performance of global search  (#2322)

+++ src/js/components/stock-movement-wizard/inbound/SendMovementPage.jsx
@@ -125,6 +125,17 @@ const SHIPMENT_FIELDS = {
       disabled: showOnly || received,
     }),
   },
+  expectedDeliveryDate: {
+    type: DateField,
+    label: 'react.stockMovement.expectedDeliveryDate.label',
+    defaultMessage: 'Expected receipt date',
+    attributes: {
+      dateFormat: 'MM/DD/YYYY',
+      required: true,
+      showTimeSelect: false,
+      autoComplete: 'off',
+    },
+  },
 };
 
 const SUPPLIER_FIELDS = {
@@ -274,6 +285,7 @@ class SendMovementPage extends Component {
       trackingNumber: values.trackingNumber || '',
       driverName: values.driverName || '',
       comments: values.comments || '',
+      expectedDeliveryDate: values.expectedDeliveryDate || '',
     };
 
     return this.saveShipment(payload);
@@ -402,6 +414,7 @@ class SendMovementPage extends Component {
             trackingNumber: stockMovementData.trackingNumber,
             driverName: stockMovementData.driverName,
             comments: stockMovementData.comments,
+            expectedDeliveryDate: stockMovementData.expectedDeliveryDate,
             // Below values are reassigned in case of editing destination or description
             name: stockMovementData.name,
             description: stockMovementData.description,
@@ -482,26 +495,29 @@ class SendMovementPage extends Component {
    * @public
    */
   sendFilesAndSave(values) {
-    this.props.showSpinner();
-    const { files } = this.state;
-    if (files.length > 1) {
-      this.sendFiles(files)
-        .then(() => {
-          Alert.success(this.props.translate('react.stockMovement.alert.filesSuccess.label', 'Files uploaded successfuly!'), { timeout: 3000 });
-          this.removeFiles(_.map(files, file => file.name));
-          this.prepareRequestAndSubmitStockMovement(values);
-        })
-        .catch(() => Alert.error(this.props.translate('react.stockMovement.alert.filesError.label', 'Error occured during files upload!')));
-    } else if (files.length === 1) {
-      this.sendFile(files[0])
-        .then(() => {
-          Alert.success(this.props.translate('react.stockMovement.alert.fileSuccess.label', 'File uploaded successfuly!'), { timeout: 3000 });
-          this.removeFile(files[0].name);
-          this.prepareRequestAndSubmitStockMovement(values);
-        })
-        .catch(() => Alert.error(this.props.translate('react.stockMovement.alert.fileError.label', 'Error occured during file upload!')));
-    } else {
-      this.prepareRequestAndSubmitStockMovement(values);
+    const errors = this.validate(values);
+    if (_.isEmpty(errors)) {
+      this.props.showSpinner();
+      const { files } = this.state;
+      if (files.length > 1) {
+        this.sendFiles(files)
+          .then(() => {
+            Alert.success(this.props.translate('react.stockMovement.alert.filesSuccess.label', 'Files uploaded successfuly!'), { timeout: 3000 });
+            this.removeFiles(_.map(files, file => file.name));
+            this.prepareRequestAndSubmitStockMovement(values);
+          })
+          .catch(() => Alert.error(this.props.translate('react.stockMovement.alert.filesError.label', 'Error occured during files upload!')));
+      } else if (files.length === 1) {
+        this.sendFile(files[0])
+          .then(() => {
+            Alert.success(this.props.translate('react.stockMovement.alert.fileSuccess.label', 'File uploaded successfuly!'), { timeout: 3000 });
+            this.removeFile(files[0].name);
+            this.prepareRequestAndSubmitStockMovement(values);
+          })
+          .catch(() => Alert.error(this.props.translate('react.stockMovement.alert.fileError.label', 'Error occured during file upload!')));
+      } else {
+        this.prepareRequestAndSubmitStockMovement(values);
+      }
     }
   }
 
@@ -512,6 +528,7 @@ class SendMovementPage extends Component {
       trackingNumber: values.trackingNumber || '',
       driverName: values.driverName || '',
       comments: values.comments || '',
+      expectedDeliveryDate: values.expectedDeliveryDate || '',
     };
 
     if ((this.props.currentLocationId !== values.origin.id) && (values.origin.type !== 'SUPPLIER' && values.hasManageInventory)) {
@@ -639,6 +656,7 @@ class SendMovementPage extends Component {
     const errors = {};
     const date = moment(this.props.minimumExpirationDate, 'MM/DD/YYYY');
     const dateShipped = moment(values.dateShipped, 'MM/DD/YYYY');
+    const expectedDeliveryDate = moment(values.expectedDeliveryDate, 'MM/DD/YYYY');
 
     if (date.diff(dateShipped) > 0) {
       errors.dateShipped = 'react.stockMovement.error.invalidDate.label';
@@ -648,6 +666,12 @@ class SendMovementPage extends Component {
     }
     if (!values.shipmentType) {
       errors.shipmentType = 'react.default.error.requiredField.label';
+    }
+    if (!values.expectedDeliveryDate) {
+      errors.expectedDeliveryDate = 'react.default.error.requiredField.label';
+    }
+    if (moment().startOf('day').diff(expectedDeliveryDate) > 0) {
+      errors.expectedDeliveryDate = 'react.stockMovement.error.pastDate.label';
     }
 
     return errors;
@@ -758,7 +782,7 @@ class SendMovementPage extends Component {
                     type="submit"
                     onClick={() => { this.sendFilesAndSave(values); }}
                     className={`${values.shipped ? 'btn btn-outline-secondary' : 'btn btn-outline-success'} float-right btn-form btn-xs`}
-                    disabled={invalid || values.statusCode === 'DISPATCHED'}
+                    disabled={values.statusCode === 'DISPATCHED'}
                   ><Translate id="react.stockMovement.sendShipment.label" defaultMessage="Send shipment" />
                   </button>
                   {values.shipped && this.props.isUserAdmin ?
